wave5 : Add v4l2 control for background detection.
Add the V4L2_CID_MPEG_VIDEO_BACKGROUND_DETECTION v4l2 control.
If enabled then, the encoder detect a background region in frame and use low bits or skip mode to encode the background region.
If a lot of scenes are stationary or background, It may help to reduce the video bitrate. Applicable to the encoder.

Signed-off-by: Nas Chung <nas.chung@chipsnmedia.com>