wave5: Avoid race condition for interrupt handling.
In case of multi instance, interrupt can be occurred for other instance as soon as interrupt is cleared.
If driver read the instance_info after clearing the interrupt, we cannot ensure the instance_info is valid for current interrupt.
Read the instance_info register for each interrupt before clear the interrupt.

Signed-off-by: Nas Chung <nas.chung@chipsnmedia.com>