wave5: Enable a full interrupt to handle a bitstream encoded more than a buffer size
In case of encoder, if a bitstream encoded is generated more than a buffer size while encoding a yuv picture,
VPU could be hang-up, so to prevent it, we have a full interrupt feature.
if a full interrupt happens when the condition happens, a driver will handle it as error.

Signed-off-by: Jackson.lee <jackson.lee@chipsnmedia.com>