media: chips-media: wave5: Fix hang issues after improving cpu load
The following is one of test cases of fluster hevc.
When we ran the following, hang happened because we check if instance_queue_count is 0 or not in the stop_streaming,
but the below stream is for testing sequence change, when sequence change happens, the instance_queue_count is always 0
until next bitstream is fed.
    === python3 ./fluster.py run -ts JCT-VC-HEVC_V1 -tv RAP_B_Bossen_2 ===
So I changed a logic to skip checking it when sequence change happens.

Signed-off-by: Jackson.lee <jackson.lee@chipsnmedia.com>